Extract ApiStream and JsonToCsvStream to separate files
